Change how typst figures are represented in the tag tree
The `Figure` element is meant to represent single images or illustations.
Generate a `Div` tag for figure elements by default.
Only use a `Figure` tag in these three cases:
- The figure element has an alternative description
- The figure element contains exactly one image
- The figure element contains only marked content
If a figure contains exactly one equation, don't emit a `Figure` tag.
If a figure contains a table, don't emit a `Figure` tag and move the
optional caption inside the table.

Always place a figure caption before the content it's describing.